+++ src/components/OfflineIndicator.js
@@ -1,24 +1,11 @@
 import { h } from 'preact'
 import { useEffect, useState } from 'preact/hooks'
-import { useI18n } from 'hooks'
+import { useI18n, useOnlineStatus } from 'hooks'
 
 export const OfflineIndicator = ({ routeUrl }) => {
   const i18n = useI18n()
-  const [isOnline, setOnline] = useState(navigator.onLine)
+  const isOnline = useOnlineStatus()
   const [isVisible, setVisible] = useState()
-
-  const onOnlineChange = () => {
-    setOnline(navigator.onLine)
-  }
-
-  useEffect(() => {
-    window.addEventListener('offline', onOnlineChange)
-    window.addEventListener('online', onOnlineChange)
-    return () => {
-      window.removeEventListener('offline', onOnlineChange)
-      window.removeEventListener('online', onOnlineChange)
-    }
-  }, [])
 
   useEffect(() => {
     setVisible(!isOnline)
