+++ src/components/Search.js
@@ -1,16 +1,35 @@
 import { h } from 'preact'
 import { useRef, useEffect } from 'preact/hooks'
 import { ListView } from 'components'
-import { useNavigation, useSearch, useI18n, useSoftkey } from 'hooks'
+import { useNavigation, useSearch, useI18n, useSoftkey, useOnlineStatus } from 'hooks'
 import { articleHistory, goto } from 'utils'
 import { getRandomArticleTitle } from 'api'
 
+const SearchOfflinePanel = () => {
+  const i18n = useI18n()
+  return (
+    <div class='search-offline-panel'>
+      <div class='search-offline-content'>
+        <img src='images/search-offline.svg' />
+        <div class='message'>{i18n.i18n('offline-message')}</div>
+      </div>
+    </div>
+  )
+}
+
 export const Search = () => {
   const containerRef = useRef()
-  const [current, setNavigation, getCurrent] = useNavigation('Search', containerRef, 'y')
+  const inputRef = useRef()
   const i18n = useI18n()
+  const [current, setNavigation, getCurrent] = useNavigation('Search', containerRef, 'y')
   const lang = i18n.locale
   const [query, setQuery, searchResults] = useSearch(lang)
+  const isOnline = useOnlineStatus(online => {
+    if (online) {
+      console.log('re setting the query', inputRef.current.value)
+      setQuery(query)
+    }
+  })
   const onKeyCenter = () => {
     const { index, key } = getCurrent()
     if (index) {
@@ -39,9 +58,10 @@ export const Search = () => {
 
   return (
     <div class='search'>
-      <img class='double-u' src='images/w.svg' style={{ display: (searchResults ? 'none' : 'block') }} />
-      <input type='text' placeholder={i18n.i18n('search-placeholder')} value={query} onInput={(e) => setQuery(e.target.value)} data-selectable />
+      <img class='double-u' src='images/w.svg' style={{ display: ((searchResults || !isOnline) ? 'none' : 'block') }} />
+      <input ref={inputRef} type='text' placeholder={i18n.i18n('search-placeholder')} value={query} onInput={(e) => setQuery(e.target.value)} data-selectable />
       { searchResults && <ListView header={i18n.i18n('header-search')} items={searchResults} containerRef={containerRef} empty={i18n.i18n('no-result-found')} /> }
+      { !isOnline && <SearchOfflinePanel /> }
     </div>
   )
 }
