+++ src/hooks/index.js
@@ -6,6 +6,7 @@ export * from './useArticleTextSize'
 export * from './useHistoryState'
 export * from './useI18n'
 export * from './useNavigation'
+export * from './useOnlineStatus'
 export * from './usePopup'
 export * from './useScroll'
 export * from './useSearch'
