+++ style/search.less
@@ -33,4 +33,20 @@
 		top: 69px;
 		bottom: @softkeysHeight;
 	}
+
+	.search-offline-panel {
+		width: 100vw;
+		height: calc( 100vh - @softkeysHeight - 41px );
+		background-color: #fff;
+		display: flex;
+
+		.search-offline-content {
+			margin: auto;
+
+			.message {
+				font-size: 17px;
+				font-family: @textFont;
+			}
+		}
+	}
 }
